Undid the nonsense with regards to the lack of reef/pages/
FUCK

This is the summation of a lot of commits. It encomasses the previous Output-Scanning branch.

Added OutputParser folder, this contains a python script that, given the path to an output folder, will return a dictionary with all the data in it.
Also added (hopefully) all the includes that are needed to read the protocol buffer. If any are missing let me know.

Removed stupid parser.py~ file. Bad GEdit.

Snapshots now into a single timeseries dict

Cleared up Elford's shit.

Fixed the folder-parsing to save file-names instead of paths

Handled the case where calling functions that did not belong to an object would cause python to break

Updated to handle the case that the path is an empty string#

Cleaned up joining of timeseries to use a map instead of a for-loop

Fixed the lack of reef/pages