Elford's work on starting and monitoring experiments
Refactored code adding an __issueCommandLineClientCommand function and added a function to start the experiment

Added the functional code to:
 * Start the experiment
 * getalloutput
 * check status of hosts

Only the last part has the appropriate handler (which functions).
Also implemented a new dictionary class to let us include Set objects
in our groups (it makes more sense in cases of lists)

Added ability to actually make requests to getalloutput, getallstatus

Added ability to start the experiment.

Added ability to do a basic status query (is the server running or not?

Created a groups class and moved it into the groups handler